chore(cleanup): split azure-devops/utils.ts into domain modules
